cleanup dead code in tests
