+++ quickwit/quickwit-datafusion/tests/metrics.rs
@@ -29,8 +29,11 @@ use quickwit_datafusion::test_utils::make_batch;
 use quickwit_datafusion::{DataFusionSessionBuilder, QuickwitObjectStoreRegistry};
 
 mod common;
+#[path = "common/metrics_splits.rs"]
+mod metrics_splits;
 
-use common::{TestSandbox, create_metrics_index, publish_split};
+use common::{TestSandbox, create_metrics_index};
+use metrics_splits::publish_split;
 
 // ── Setup ──────────────────────────────────────────────────────────
